+++ BitFaster.Caching/Lru/ConcurrentLruBuilder.cs
@@ -52,36 +52,4 @@ namespace BitFaster.Caching.Lru
             }
         }
     }
-
-    public sealed class AsyncConcurrentLruBuilder<K, V> : LruBuilderBase<K, V, AsyncConcurrentLruBuilder<K, V>, IAsyncCache<K, V>>
-    {
-        /// <summary>
-        /// Creates a ConcurrentLruBuilder.
-        /// </summary>
-        public AsyncConcurrentLruBuilder()
-            : base(new LruInfo<K>())
-        {
-        }
-
-        internal AsyncConcurrentLruBuilder(LruInfo<K> info)
-            : base(info)
-        {
-        }
-
-        ///<inheritdoc/>
-        public override IAsyncCache<K, V> Build()
-        {
-            switch (info)
-            {
-                case LruInfo<K> i when i.WithMetrics && !i.TimeToExpireAfterWrite.HasValue:
-                    return new ConcurrentLru<K, V>(info.ConcurrencyLevel, info.Capacity, info.KeyComparer);
-                case LruInfo<K> i when i.WithMetrics && i.TimeToExpireAfterWrite.HasValue:
-                    return new ConcurrentTLru<K, V>(info.ConcurrencyLevel, info.Capacity, info.KeyComparer, info.TimeToExpireAfterWrite.Value);
-                case LruInfo<K> i when i.TimeToExpireAfterWrite.HasValue:
-                    return new FastConcurrentTLru<K, V>(info.ConcurrencyLevel, info.Capacity, info.KeyComparer, info.TimeToExpireAfterWrite.Value);
-                default:
-                    return new FastConcurrentLru<K, V>(info.ConcurrencyLevel, info.Capacity, info.KeyComparer);
-            }
-        }
-    }
 }
